Resolve debt register clear-out
