+++ docs/_archive/debt/debt-428-question-ids-narrowed-unverified-against-legacy-data.md
@@ -1,6 +1,6 @@
 # DEBT-428: questionIds validation narrowed to strict zUuid + count-match without auditing existing data
 
-**Status:** Implemented; post-deploy proof pending
+**Status:** Resolved
 **Priority:** P3
 **Date:** 2026-06-30
 
@@ -8,13 +8,13 @@
 
 ## Description
 
-`practice-session-params.ts`'s `questionIds` field was tightened from `z.array(z.string().min(1))` to `z.array(zUuid).min(1)` plus a new `count === questionIds.length` refine, applied via `parsePracticeSessionParamsJson(row.paramsJson, 'INTERNAL_ERROR')` on every read of every `practice_sessions` row (`findByIdAndUserId`, `findLatestIncompleteByUserId`, `findCompletedByUserId`, all funneling through the repository's private `toDomain`). Two additional strict-parse sites exist as of 2026-07-04: `create()` validates params on write (`drizzle-practice-session-repository.ts:296`), and the missing-row classifier in `practice-session-question-state-updater.ts:85` parses params when a state row is absent.
+`practice-session-params.ts`'s `questionIds` field was tightened from `z.array(z.string().min(1))` to `z.array(zUuid).min(1)` plus a new `count === questionIds.length` refine, applied via `parsePracticeSessionParamsJson(row.paramsJson, 'INTERNAL_ERROR')` on every read of every `practice_sessions` row (`findByIdAndUserId`, `findLatestIncompleteByUserId`, `findCompletedByUserId`, all funneling through the repository's private `toDomain`). Two additional strict-parse sites exist as of 2026-07-04: `create()` validates params on write (`drizzle-practice-session-repository.ts:284`), and the missing-row classifier in `practice-session-question-state-updater.ts:85` parses params when a state row is absent.
 
 ## Impact
 
 Any pre-existing row that doesn't conform to the new stricter shape (a non-UUID-format legacy `questionId`, an empty `questionIds` array, or a historical `count`/`questionIds.length` mismatch from an old bug) now throws `INTERNAL_ERROR` on read, where the old looser schema parsed it successfully.
 
-Post-deploy proof on 2026-07-02 turned this from unverified risk into concrete Development cleanup: Production has zero affected rows, but Development has one ended tutor session whose `params_json` is a double-encoded JSON string rather than a top-level object. A follow-up read-only inspection on 2026-07-02 identified it as session `00000000-0000-4000-8000-000000000244` for user `3955ed2e-9034-4943-ad15-04cd79daa568`, `mode='tutor'`, `started_at='2026-01-01T00:00:00.000Z'`, `ended_at='2026-01-01T00:02:00.000Z'`. It has zero normalized `practice_session_question_states` rows and one linked attempt row. The row therefore has no top-level `questionIds` as far as JSONB operators and the current parser are concerned, and the embedded string contains legacy malformed question IDs. Object-shaped sessions in both environments have zero duplicate and zero dangling `questionIds`.
+Post-deploy proof on 2026-07-02 turned this from unverified risk into concrete Development cleanup: Production had zero affected rows, but Development had one ended tutor session whose `params_json` was a double-encoded JSON string rather than a top-level object. A follow-up read-only inspection on 2026-07-02 identified it as session `00000000-0000-4000-8000-000000000244` for user `3955ed2e-9034-4943-ad15-04cd79daa568`, `mode='tutor'`, `started_at='2026-01-01T00:00:00.000Z'`, `ended_at='2026-01-01T00:02:00.000Z'`. It had zero normalized `practice_session_question_states` rows and one linked attempt row. The row therefore had no top-level `questionIds` as far as JSONB operators and the current parser were concerned until `0026_track_a_tail_sweep` normalized the string payload back to an object. Object-shaped sessions in both environments had zero duplicate and zero dangling `questionIds`.
 
 **Correction (2026-06-30, second review pass):** the `questionIds` *uniqueness* refine (`new Set(questionIds).size === questionIds.length`) is **not** new — it was already present and enforced on `main` before this PR; this PR only added the `zUuid` format check, `.min(1)`, and the `count === questionIds.length` refine. A legacy row with duplicate `questionIds` would already have failed to parse before this PR too, so uniqueness itself is not a newly introduced risk here. What *is* worth tracking is a downstream interaction: migration `0021`'s backfill `INSERT` uses `ON CONFLICT (practice_session_id, question_id) DO NOTHING`, so a row with duplicate `questionIds` (if one somehow exists despite the app-level guard, e.g. from a pre-Zod-refine era or a direct DB write) would backfill fewer state rows than `questionIds.length`, producing a `rows.length < params.questionIds.length` mismatch at read time — a different, migration-level manifestation of the same underlying data-quality question. Tracked as an explicit scenario under [DEBT-427](./debt-427-migration-fk-ordering-and-unaudited-cleanup.md) rather than duplicated here.
 
@@ -22,7 +22,7 @@ Post-deploy proof on 2026-07-02 turned this from unverified risk into concrete D
 
 Implemented in migration `0026_track_a_tail_sweep.sql`: string-typed `practice_sessions.params_json` rows are converted back into JSONB objects by parsing the string payload. The migration fails loudly if any string row cannot be parsed to a JSON object; it does not add application parser tolerance for double-encoded JSON. This keeps the invariant at the data boundary and avoids making internal Development residue a permanent compatibility surface.
 
-The same migration also records the affected row count with `GET DIAGNOSTICS` / `RAISE NOTICE`, establishing the row-count audit pattern requested by DEBT-427 for future cleanup migrations. Production previously had zero affected rows; this debt remains active only until post-merge deploy proof confirms both Neon branches are at ledger head `0026_track_a_tail_sweep` and have zero string-typed `params_json` rows.
+The same migration also records the affected row count with `GET DIAGNOSTICS` / `RAISE NOTICE`, establishing the row-count audit pattern requested by DEBT-427 for future cleanup migrations. Production previously had zero affected rows. Post-merge proof on 2026-07-04 confirmed both deployed Neon branches are at ledger head `0026_track_a_tail_sweep` and have zero string-typed `params_json` rows, so no application parser tolerance was added and this debt is closed.
 
 ## Verification
 
@@ -37,7 +37,14 @@ Implementation proof in this branch:
 
 - `tests/integration/practice-session-params-json-cleanup.integration.test.ts` seeds a string-typed `params_json` row, runs the marked `DEBT-428/434` cleanup block twice, and proves the row parses through `parsePracticeSessionParamsJson()` after cleanup.
 
-Post-merge proof still required before archive:
+Post-merge proof recorded 2026-07-04:
+
+| Target | Host | Ledger entries | Ledger head | String-typed `params_json` rows | Sanity row `00000000-0000-4000-8000-000000000244` |
+|---|---|---:|---|---:|---|
+| Development | `ep-still-frog` | 27 | `0026_track_a_tail_sweep` / `1783092600000` | 0 | Parses through `parsePracticeSessionParamsJson()` as an object with `count=2`, `questionIds.length=2` |
+| Production | `ep-withered-cell` | 27 | `0026_track_a_tail_sweep` / `1783092600000` | 0 | Absent |
+
+Query:
 
 ```sql
 SELECT count(*)::int AS string_params_json_rows
@@ -45,9 +52,7 @@ FROM practice_sessions
 WHERE jsonb_typeof(params_json) = 'string';
 ```
 
-Expected result for Development (`ep-still-frog`) and Production (`ep-withered-cell`): `0`.
-
 ## Related
 
-- PR #537, [DEBT-425](../_archive/debt/debt-425-legacy-compatibility-tolerances-audit.md)
+- PR #537, [DEBT-425](./debt-425-legacy-compatibility-tolerances-audit.md)
 - `src/adapters/repositories/practice-session-params.ts:26-45` (`questionIds` schema block; shifted from 24-32 by the DEBT-433 cross-reference comment added in PR #556)
