+++ docs/dev/migration-authoring.md
@@ -1,0 +1,104 @@
+# Migration Authoring
+
+**Last Updated:** 2026-07-04
+
+This guide is the durable review checklist for Drizzle migration files. It
+complements [Deployment Procedure](./deployment-procedure.md), which describes
+when migrations run. This file describes how to write migrations so data cleanup,
+constraints, and lock scope are explicit before a PR merges.
+
+---
+
+## Pre-Flight Data Proof
+
+Before adding a migration that tightens constraints, backfills derived rows, or
+changes a JSON/data invariant, run read-only proof queries against the target
+shape of the current data.
+
+At minimum, count the rows that could fail the migration:
+
+- dangling references the new foreign key will reject;
+- duplicate keys the new unique constraint will reject;
+- malformed JSON/data shapes a parser or cast will reject;
+- rows a cleanup statement intends to mutate.
+
+Record the counts in the bug/debt/spec document that justifies the migration.
+The migration should then either fail loudly when the precondition is violated
+or include a deterministic cleanup before the constraint that depends on it.
+
+## Operation Ordering
+
+Backfills and cleanups that make data satisfy a constraint must run before the
+constraint is added. Do not add a foreign key, unique constraint, or check
+constraint first and hope the later backfill explains the failure.
+
+`0021_flaky_domino.sql` is the documented caution: it added foreign keys before
+its own legacy-data backfill. That had already applied successfully in the
+deployed branches, but the ordering is not the authoring pattern to repeat.
+
+Safer examples are the cleanup-before-constraint steps in `0022`, `0023`, and
+`0024`; they normalize inconsistent data before adding the constraints that
+would otherwise reject it.
+
+## Cleanup Audit Trail
+
+Every data-mutating cleanup statement in a migration should emit an affected-row
+count. For hand-authored SQL, use a `DO` block with `GET DIAGNOSTICS ... =
+ROW_COUNT` and `RAISE NOTICE` for each cleanup.
+
+`0026_track_a_tail_sweep.sql` is the local exemplar:
+
+- it marks the cleanup block;
+- it fails loudly if string-typed `params_json` cannot parse to an object;
+- it records the count of normalized string rows;
+- it records the count of stripped stale `params_json.questionStates` keys.
+
+Those notices make deploy logs useful months later, when the old state of the
+table can no longer be reconstructed from the migration ledger alone.
+
+## Lock Scope
+
+Drizzle applies pending migration files in a transaction. PostgreSQL locks taken
+inside that transaction are held until the transaction commits, not merely until
+the statement finishes.
+
+Review lock scope before merging migrations that touch live tables:
+
+- `ALTER TABLE ... ADD CONSTRAINT ... FOREIGN KEY` takes
+  `SHARE ROW EXCLUSIVE` locks on the referencing and referenced tables while the
+  constraint is added and validated. This blocks concurrent writes.
+- Standard `CREATE INDEX` takes a `SHARE` lock. It blocks concurrent writes but
+  not ordinary reads.
+- `CREATE INDEX CONCURRENTLY` cannot run inside a transaction block. If a large
+  live table needs a write-nonblocking index build, split it into a
+  separately-applied migration path rather than putting it in the normal Drizzle
+  transaction batch.
+
+Small or empty tables can accept the standard lock cost, but the migration must
+say so in a nearby SQL comment. `0026_track_a_tail_sweep.sql` does this for its
+bounded cleanup over the small `practice_sessions` table.
+
+## Deployment Contract
+
+Never use `drizzle-kit push` for this repository. It bypasses checked-in
+migration files and can miss extensions, constraints, custom SQL, and ledger
+history.
+
+Git-triggered Vercel Preview and Production deploys run the checked-in migration
+journal through the configured Build Command:
+
+```bash
+pnpm db:migrate && pnpm build
+```
+
+For any manual fallback or operator-run migration, pass the intended database
+explicitly:
+
+```bash
+DATABASE_URL="<verified target connection string>" pnpm db:migrate
+```
+
+Verify the target host without printing credentials before mutating any remote
+database. CI migration success only proves the journal applies to CI's
+throwaway Postgres service; it does not migrate Vercel Preview or Production
+databases.
