+++ src/adapters/repositories/practice-session-question-state-updater.ts
@@ -20,7 +20,7 @@ type QuestionStateSnapshot = {
   endedAt: Date | null;
 };
 
-function toDomainQuestionState(
+export function toDomainQuestionState(
   row: PracticeSessionQuestionStateRow,
 ): PracticeSessionQuestionState {
   return {
